services/mcp-content-generator/requirements.txt
fastmcp==0.1.0
anthropic>=0.34.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
